--- daren-app/src/components/Card/Card.css
@@ -1,30 +1,0 @@
-.card {
-    position: relative;
-    width: 300px;
-    height: 250px;
-    border-radius: 8px;
-    overflow: hidden;
-  }
-
-  .card-image {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-
- .card-overlay {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    padding: 10px;
-    box-sizing: border-box;
-    background-color: rgba(0, 0, 0, 0.7);
-    color: #fff;
-    font-size: 16px;
-    line-height: 1.4;
-  }
-.image-grayscale{
-  filter: grayscale(100%);
-  
-}
