+++ daren-app/src/components/Header/Header.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import './Header.css'
 import HeaderText from '../../components/HeaderText/HeaderText'
-import Switcher from '../../components/Switcher/Switcher'
-import SecondaryHeader from '../../components/SecondaryHeaderText/SecondaryHeaderText'
-import Button from '../../components/Button/Button'
+import Switcher from '../UI/Switcher/Switcher'
+import SecondaryHeader from "../SecondaryHeaderText/SecondaryHeaderText"
+import Button from "../UI/Button/Button"
 import Hgroup from '../../components/Hgroup/Hgroup'
 
 const Header = () => {
