+++ daren-app/src/components/Main/Main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import "./Main.css"
-import Card from '../Card/Card'
+import Card from '../UI/Card/Card'
 import Image from '../../assests/images/img.jpg'
 import { useGlobalContext } from '../../context/context'
