+++ daren-app/src/components/UI/Card/Card.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import "./Card.css"
+
+const Card = ({id , img, description, switcher}) => {
+  return (
+    <>
+    <article key={id} class="card">
+     <img className={`card-image ${switcher?'image-grayscale': ''}`} src={img} alt="Background Image"/>
+      <figcaption class="card-overlay">
+         {description}
+     </figcaption>
+    </article>
+    </>
+  )
+}
+
+export default Card
